app-manager extract GitOperations
